Added code_of_conduct.html & health_policy.html
Suggested by https://chaos.social/@dentangle/112156697328041936

Not putting in placeholders as this may encourage events to leave these in place
when not actually intending to use them and not being willing to put the work in to make them effective.

Did put in comment to link to one in template form.